+++ cardano-slotting/CHANGELOG.md
@@ -1,3 +1,10 @@
+# Changelog for `cardano-slotting`
+
+## 0.1.1.0
+
+* Addition of `ToJSON`/`FromJSON` instances for `SystemStart`
+* Addition of `ToCBOR`/`FromCBOR` instances for `RelativeTime` and `SlotLength`
+
 ## 0.1.0.1
 
 * Initial release